cleanup; docs
